+++ .env.onepack
@@ -1,0 +1,2 @@
+VITE_BUILD_TYPE = Production
+VITE_BUILD_PLATFORM = OnePack
